Decouple collision layer to module.
