+++ public/js/layers/collision.js
@@ -1,0 +1,37 @@
+export function createCollisionLayer(level) {
+    const resolvedTiles = [];
+
+    const tileResolver = level.tileCollider.tiles;
+    const tileSize = tileResolver.tileSize;
+
+    const getByIndexOriginal = tileResolver.getByIndex;
+    tileResolver.getByIndex = function getByIndexFake(x, y) {
+        resolvedTiles.push({x, y});
+        return getByIndexOriginal.call(tileResolver, x, y);
+    }
+
+    return function drawCollision(context, camera) {
+        context.strokeStyle = 'blue';
+        resolvedTiles.forEach(({x, y}) => {
+            context.beginPath();
+            context.rect(
+                x * tileSize - camera.pos.x,
+                y * tileSize - camera.pos.y,
+                tileSize, tileSize);
+            context.stroke();
+        });
+
+        context.strokeStyle = 'red';
+        level.entities.forEach(entity => {
+            context.beginPath();
+            context.rect(
+                entity.bounds.left - camera.pos.x,
+                entity.bounds.top - camera.pos.y,
+                entity.size.x,
+                entity.size.y);
+            context.stroke();
+        });
+
+        resolvedTiles.length = 0;
+    };
+}
